Claim: Write a getter for the `label` property
The value of this only depends on the version, so this shouldn't
be a property that can be written directly. Anyhow, this field
was never written and remained `undefined`. As a side effect of
this change, the claim now also turns up as entry when filling
in the `componentStore` field of the manifest.